% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_only.R
\name{keep_only}
\alias{keep_only}
\title{只保留指定对象，删除全局环境中的其他对象}
\usage{
keep_only(keep_names, envir = .GlobalEnv)
}
\arguments{
\item{keep_names}{字符向量，要保留的对象名。}

\item{envir}{环境，默认 \code{.GlobalEnv}。}
}
\value{
无返回值（执行环境修改）。
}
\description{
从全局环境中删除除指定对象外的所有对象。
}
\examples{
keep_only(c("d", "df"))
keep_only(c("d", "df"))
ls()  # 只剩 d 和 df
}
