% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_df_vars.R
\name{print.inspect_df_vars}
\alias{print.inspect_df_vars}
\title{Pretty print for \code{inspect_df_vars}}
\usage{
\method{print}{inspect_df_vars}(x, ...)
}
\arguments{
\item{x}{\code{inspect_df_vars} 对象。}

\item{...}{保留参数。}
}
\description{
竖向块状打印，每个变量单独一组，阅读更友好。
}
