% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_and_count.R
\name{remove_na_and_count}
\alias{remove_na_and_count}
\title{删除指定列中的NA并统计删除行数}
\usage{
remove_na_and_count(data, col_name)
}
\arguments{
\item{data}{数据框}

\item{col_name}{列名（字符型）}
}
\value{
删除NA后的数据框
}
\description{
删除数据框中指定列的NA值，并返回删除后的数据及删除信息
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c("x", "y", "z"))
remove_na_and_count(df, "a")
}
