% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_fread.R
\name{z_fread}
\alias{z_fread}
\title{z_fread: Fast and robust CSV/TSV reader (wrapper of data.table::fread)}
\usage{
z_fread(
  path,
  sep = ",",
  encoding = "UTF-8",
  na = c("", "NA", "NaN"),
  select = NULL,
  drop = NULL,
  n_max = NULL,
  as_data_frame = TRUE,
  show_progress = FALSE
)
}
\description{
z_fread: Fast and robust CSV/TSV reader (wrapper of data.table::fread)
}
